+++ readPE.h
@@ -13,4 +13,4 @@ VOID readDosHeader(HWND hwnd, LPVOID pFileBuffer);
 
 VOID readFileHeader(HWND hwnd, LPVOID pFileBuffer);
 
-VOID writeToText(HWND hwnd, INT TEXT_ID, CONST TCHAR* format, WORD data);
+VOID writeToText(HWND hwnd, INT TEXT_ID, CONST TCHAR* format, DWORD data);
